fix: 안쓰는 연관관계 및 파일 삭제
