+++ server/src/artwork/artwork.entity.ts
@@ -2,8 +2,6 @@ import { Column, Entity, ManyToOne, OneToMany, OneToOne, PrimaryGeneratedColumn 
 import { User } from '../user/user.entity';
 import { Auction } from '../auction/auction.entity';
 import { ArtworkStatus } from './artwork.status.enum';
-import { InterestArtwork } from '../interestArtwork/interestArtwork.entity';
-import { ArtworkInBid } from '../artworkInBid/artworkInBid.entity';
 
 @Entity()
 export class Artwork {
@@ -47,17 +45,11 @@ export class Artwork {
     @Column({ nullable: true })
     exhibitionId: number;
 
-    @ManyToOne(type => User, user => user.artworks)
+    @ManyToOne(type => User)
     artist: User;
 
-    @ManyToOne(type => User, user => user.biddedArtworks)
+    @ManyToOne(type => User)
     owner: User;
-
-    @OneToMany(type => InterestArtwork, interestArtwork => interestArtwork.artwork)
-    interestedUsers: InterestArtwork[];
-
-    @OneToMany(type => ArtworkInBid, artworkInBid => artworkInBid.artwork)
-    usersInBid: User[];
 
     @OneToOne(type => Auction, auction => auction.artwork, {
         cascade: true,
