+++ server/src/artwork/artwork.module.ts
@@ -3,8 +3,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuctionRepository } from 'src/auction/auction.repository';
 import { ImageModule } from 'src/image/image.module';
 import { ArtworkRepository } from './artwork.repository';
-import { ArtworkController } from './controller/artwork.controller';
-import { ArtworkService } from './service/artwork.service';
+import { ArtworkController } from './artwork.controller';
+import { ArtworkService } from './artwork.service';
 import { UserRepository } from '../user/user.repository';
 import { JwtModule } from '@nestjs/jwt';
 import { InterestArtworkModule } from '../interestArtwork/interestArtwork.module';
