+++ server/src/auction/auction.gateway.ts
@@ -9,7 +9,7 @@ import {
 } from '@nestjs/websockets';
 import { Server, Socket } from 'socket.io';
 import { AuctionHistoryService } from '../auctionHistory/auctionHistory.service';
-import AuctionService from './service/auction.service';
+import AuctionService from './auction.service';
 
 @WebSocketGateway({
     namespace: '/auction',
