+++ server/src/auction/auction.module.ts
@@ -3,8 +3,8 @@ import { AuctionGateway } from './auction.gateway';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuctionRepository } from './auction.repository';
 import { AuctionHistoryModule } from '../auctionHistory/auctionHistory.module';
-import AuctionController from './controller/auction.controller';
-import AuctionService from './service/auction.service';
+import AuctionController from './auction.controller';
+import AuctionService from './auction.service';
 
 @Module({
     imports: [AuctionHistoryModule, TypeOrmModule.forFeature([AuctionRepository])],
