+++ server/src/auth/auth.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
-import { UserService } from '../../user/service/user.service';
+import { UserService } from '../user/user.service';
 import { JwtService } from '@nestjs/jwt';
-import { User } from '../../user/user.entity';
+import { User } from '../user/user.entity';
 import { UpdateResult } from 'typeorm';
 import axios from 'axios';
