+++ server/src/cronTask/crontask.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
-import { ArtworkService } from 'src/artwork/service/artwork.service';
-import AuctionService from 'src/auction/service/auction.service';
+import { ArtworkService } from 'src/artwork/artwork.service';
+import AuctionService from 'src/auction/auction.service';
 import { ethers } from 'ethers';
 import * as contractAddress from './ethereum/address.json';
 import * as abi from './ethereum/abi.json';
