+++ server/src/exhibition/exhibition.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ExhibitionController } from './controller/exhibition.controller';
-import { ExhibitionService } from './service/exhibition.service';
+import { ExhibitionController } from './exhibition.controller';
+import { ExhibitionService } from './exhibition.service';
 import { ExhibitionRepository } from './exhibition.repository';
 import { JwtModule } from '@nestjs/jwt';
-import { ImageService } from 'src/image/service/image.service';
+import { ImageService } from 'src/image/image.service';
 import { UserRepository } from 'src/user/user.repository';
 import { ArtworkRepository } from 'src/artwork/artwork.repository';
