+++ server/src/interestArtwork/interestArtwork.entity.ts
@@ -1,5 +1,4 @@
-import { Entity,   ManyToOne, PrimaryGeneratedColumn} from 'typeorm';
-import { User } from '../user/user.entity';
+import { Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { Artwork } from '../artwork/artwork.entity';
 
 @Entity()
@@ -7,9 +6,8 @@ export class InterestArtwork {
     @PrimaryGeneratedColumn()
     id: number;
 
-    @ManyToOne(type => User, user => user.interestArtworks)
-    user: User;
+    userId: number;
 
-    @ManyToOne(type => Artwork, artwork => artwork.interestedUsers)
+    @ManyToOne(type => Artwork)
     artwork: Artwork;
 }
