+++ server/src/user/user.controller.ts
@@ -1,10 +1,10 @@
 import { Body, Controller, Get, Put, Query, Req, UploadedFile, UseGuards, UseInterceptors } from '@nestjs/common';
-import { UserService } from '../service/user.service';
-import { User } from '../user.entity';
-import { Artwork } from '../../artwork/artwork.entity';
-import { Exhibition } from '../../exhibition/exhibition.entity';
-import { CustomAuthGuard } from '../../auth/guard/CustomAuthGuard';
-import { RequestUserDTO } from '../dto/userDTO';
+import { UserService } from './user.service';
+import { User } from './user.entity';
+import { Artwork } from '../artwork/artwork.entity';
+import { Exhibition } from '../exhibition/exhibition.entity';
+import { CustomAuthGuard } from '../auth/guard/CustomAuthGuard';
+import { RequestUserDTO } from './dto/userDTO';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { ApiBody, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { UpdateResult } from 'typeorm';
@@ -17,7 +17,7 @@ import {
     getBiddedArtworksApiOperation,
     getUsersExhibitionsApiOperation,
     updateUserProfileApiBody,
-} from '../swagger';
+} from './swagger';
 
 @UseGuards(CustomAuthGuard)
 @Controller('/users')
